refactor : 주 키의 타입이 Long 임으로 변경

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/service/LikeablePersonService.java
@@ -51,7 +51,7 @@ public class LikeablePersonService {
     }
 
     @Transactional
-    public RsData<LikeablePerson> delete(Integer id){
+    public RsData<LikeablePerson> delete(Long id){
         Optional<LikeablePerson> likeablePerson = likeablePersonRepository.findById(id);
 
         if(likeablePerson.isEmpty()){
